+++ backend/integration_tests/src/p2p_swap_tests.rs
@@ -1,13 +1,12 @@
 use crate::env::ENV;
 use crate::rng::{random_message_id, random_string};
 use crate::utils::tick_many;
-use crate::{client, CanisterIds, TestEnv, User};
+use crate::{client, TestEnv};
 use candid::Principal;
-use pocket_ic::PocketIc;
 use std::ops::Deref;
 use std::time::Duration;
 use test_case::test_case;
-use types::{ChatEvent, ChatId, Cryptocurrency, MessageContent, MessageContentInitial, P2PSwapContentInitial, P2PSwapStatus};
+use types::{ChatEvent, Cryptocurrency, MessageContent, MessageContentInitial, P2PSwapContentInitial, P2PSwapStatus};
 use utils::time::{DAY_IN_MS, MINUTE_IN_MS};
 
 #[test]
@@ -20,7 +19,8 @@ fn p2p_swap_in_direct_chat_succeeds() {
         ..
     } = wrapper.env();
 
-    let TestData { user1, user2, .. } = init_test_data(env, canister_ids, *controller, true);
+    let user1 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
 
     client::icrc1::happy_path::transfer(
         env,
@@ -131,7 +131,11 @@ fn p2p_swap_in_group_succeeds() {
         ..
     } = wrapper.env();
 
-    let TestData { user1, user2, group_id } = init_test_data(env, canister_ids, *controller, true);
+    let user1 = client::register_diamond_user(env, canister_ids, *controller);
+    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+
+    let group_id = client::user::happy_path::create_group(env, &user1, &random_string(), true, true);
+    client::local_user_index::happy_path::join_group(env, user2.principal, canister_ids.local_user_index, group_id);
 
     client::icrc1::happy_path::transfer(
         env,
@@ -216,7 +220,7 @@ fn p2p_swap_in_group_succeeds() {
 
     if let ChatEvent::Message(m) = event {
         if let MessageContent::P2PSwap(p) = m.content {
-            assert!(matches!(p.status, P2PSwapStatus::Accepted(c) if c.accepted_by == user2.user_id));
+            assert!(matches!(p.status, P2PSwapStatus::Completed(c) if c.accepted_by == user2.user_id));
         } else {
             panic!();
         }
@@ -236,7 +240,8 @@ fn cancel_p2p_swap_in_direct_chat_succeeds(delete_message: bool) {
         ..
     } = wrapper.env();
 
-    let TestData { user1, user2, .. } = init_test_data(env, canister_ids, *controller, true);
+    let user1 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
 
     let original_chat_balance = 11_000_000_000;
 
@@ -320,6 +325,150 @@ fn cancel_p2p_swap_in_direct_chat_succeeds(delete_message: bool) {
         client::icrc1::happy_path::balance_of(env, canister_ids.chat_ledger, Principal::from(user1.user_id)),
         original_chat_balance - (2 * Cryptocurrency::CHAT.fee().unwrap())
     );
+
+    if !delete_message {
+        let user1_event = client::user::happy_path::events_by_index(env, &user1, user2.user_id, vec![0.into()])
+            .events
+            .pop()
+            .unwrap()
+            .event;
+
+        if let ChatEvent::Message(m) = user1_event {
+            if let MessageContent::P2PSwap(p) = m.content {
+                assert!(matches!(p.status, P2PSwapStatus::Cancelled(c) if c.token0_txn_out.is_some()));
+            }
+        }
+
+        let user2_event = client::user::happy_path::events_by_index(env, &user2, user1.user_id, vec![0.into()])
+            .events
+            .pop()
+            .unwrap()
+            .event;
+
+        if let ChatEvent::Message(m) = user2_event {
+            if let MessageContent::P2PSwap(p) = m.content {
+                assert!(matches!(p.status, P2PSwapStatus::Cancelled(c) if c.token0_txn_out.is_some()));
+            }
+        }
+    }
+}
+
+#[test_case(true)]
+#[test_case(false)]
+fn cancel_p2p_swap_in_group_chat_succeeds(delete_message: bool) {
+    let mut wrapper = ENV.deref().get();
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+        ..
+    } = wrapper.env();
+
+    let user1 = client::register_diamond_user(env, canister_ids, *controller);
+    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+
+    let group_id = client::user::happy_path::create_group(env, &user1, &random_string(), true, true);
+    client::local_user_index::happy_path::join_group(env, user2.principal, canister_ids.local_user_index, group_id);
+
+    let original_chat_balance = 11_000_000_000;
+
+    client::icrc1::happy_path::transfer(
+        env,
+        *controller,
+        canister_ids.chat_ledger,
+        Principal::from(user1.user_id),
+        original_chat_balance,
+    );
+
+    let message_id = random_message_id();
+
+    let send_message_response = client::user::send_message_with_transfer_to_group(
+        env,
+        user1.principal,
+        user1.canister(),
+        &user_canister::send_message_with_transfer_to_group::Args {
+            group_id,
+            thread_root_message_index: None,
+            message_id,
+            content: MessageContentInitial::P2PSwap(P2PSwapContentInitial {
+                token0: Cryptocurrency::CHAT.try_into().unwrap(),
+                token0_amount: 10_000_000_000,
+                token1: Cryptocurrency::InternetComputer.try_into().unwrap(),
+                token1_amount: 1_000_000_000,
+                expires_in: DAY_IN_MS,
+                caption: None,
+            }),
+            sender_name: user1.username(),
+            sender_display_name: None,
+            replies_to: None,
+            mentioned: Vec::new(),
+            correlation_id: 0,
+            rules_accepted: None,
+            message_filter_failed: None,
+        },
+    );
+
+    assert!(matches!(
+        send_message_response,
+        user_canister::send_message_with_transfer_to_group::Response::Success(_)
+    ));
+
+    if delete_message {
+        let delete_message_response = client::group::delete_messages(
+            env,
+            user1.principal,
+            group_id.into(),
+            &group_canister::delete_messages::Args {
+                thread_root_message_index: None,
+                message_ids: vec![message_id],
+                as_platform_moderator: None,
+                correlation_id: 0,
+            },
+        );
+
+        assert!(matches!(
+            delete_message_response,
+            group_canister::delete_messages::Response::Success
+        ));
+
+        env.advance_time(Duration::from_millis(5 * MINUTE_IN_MS));
+    } else {
+        let cancel_swap_response = client::group::cancel_p2p_swap(
+            env,
+            user1.principal,
+            group_id.into(),
+            &group_canister::cancel_p2p_swap::Args {
+                thread_root_message_index: None,
+                message_id,
+            },
+        );
+
+        assert!(matches!(
+            cancel_swap_response,
+            group_canister::cancel_p2p_swap::Response::Success
+        ));
+    }
+
+    tick_many(env, 5);
+
+    assert_eq!(
+        client::icrc1::happy_path::balance_of(env, canister_ids.chat_ledger, Principal::from(user1.user_id)),
+        original_chat_balance - (2 * Cryptocurrency::CHAT.fee().unwrap())
+    );
+
+    if !delete_message {
+        let event = client::group::happy_path::events_by_index(env, &user1, group_id, vec![2.into()])
+            .events
+            .pop()
+            .unwrap()
+            .event;
+
+        if let ChatEvent::Message(m) = event {
+            if let MessageContent::P2PSwap(p) = m.content {
+                assert!(matches!(p.status, P2PSwapStatus::Cancelled(c) if c.token0_txn_out.is_some()));
+            }
+        }
+    }
 }
 
 #[test]
@@ -332,7 +481,8 @@ fn deposit_refunded_if_swap_expires() {
         ..
     } = wrapper.env();
 
-    let TestData { user1, user2, .. } = init_test_data(env, canister_ids, *controller, true);
+    let user1 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
 
     let original_chat_balance = 11_000_000_000;
 
@@ -390,7 +540,7 @@ fn deposit_refunded_if_swap_expires() {
 
     if let ChatEvent::Message(m) = user1_event {
         if let MessageContent::P2PSwap(p) = m.content {
-            assert!(matches!(p.status, P2PSwapStatus::Expired(c) if c.token0_txn_out.is_some()));
+            assert!(matches!(p.status, P2PSwapStatus::Expired(e) if e.token0_txn_out.is_some()));
         }
     }
 
@@ -402,25 +552,7 @@ fn deposit_refunded_if_swap_expires() {
 
     if let ChatEvent::Message(m) = user2_event {
         if let MessageContent::P2PSwap(p) = m.content {
-            assert!(matches!(p.status, P2PSwapStatus::Expired(c) if c.token0_txn_out.is_some()));
+            assert!(matches!(p.status, P2PSwapStatus::Expired(e) if e.token0_txn_out.is_some()));
         }
     }
-}
-
-fn init_test_data(env: &mut PocketIc, canister_ids: &CanisterIds, controller: Principal, public: bool) -> TestData {
-    let user1 = client::register_diamond_user(env, canister_ids, controller);
-    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
-
-    let group_name = random_string();
-
-    let group_id = client::user::happy_path::create_group(env, &user1, &group_name, public, true);
-    client::local_user_index::happy_path::join_group(env, user2.principal, canister_ids.local_user_index, group_id);
-
-    TestData { user1, user2, group_id }
-}
-
-struct TestData {
-    user1: User,
-    user2: User,
-    group_id: ChatId,
 }