+++ frontend/app/src/components/EditLabel.svelte
@@ -9,27 +9,44 @@
     import TextArea from "./TextArea.svelte";
     import { getContext } from "svelte";
     import { OpenChat } from "openchat-client";
+    import ErrorMessage from "./ErrorMessage.svelte";
+    import { toastStore } from "../stores/toast";
 
     const client = getContext<OpenChat>("client");
 
     let busy = false;
     let suggestion = "";
+    let saved = false;
 
     $: yourLanguage = supportedLanguages.find((l) => l.code === $locale)?.name ?? "English";
-    $: corrections = client.translationCorrectionsStore;
-    $: userStore = client.userStore;
+    $: englishValue = $editingLabel && $_($editingLabel.key, { locale: "en" });
+    $: englishTokens = extractTokens(englishValue);
+    $: tokenMismatch = !tokensMatch(suggestion, englishTokens);
+    $: valid = suggestion !== "" && !tokenMismatch;
 
-    $: console.log("Corrections: ", $corrections);
+    function tokensMatch(suggestion: string, originalTokens: Set<string>): boolean {
+        return setsAreEqual(extractTokens(suggestion), originalTokens);
+    }
 
-    $: existingCorrection =
-        $locale && $editingLabel && $corrections[$locale]
-            ? $corrections[$locale][$editingLabel.key]
-            : undefined;
-    $: correctedBy =
-        existingCorrection !== undefined ? $userStore[existingCorrection?.proposedBy].username : "";
+    function setsAreEqual(a: Set<string>, b: Set<string>) {
+        return [...a].every((x) => b.has(x)) && [...b].every((x) => a.has(x));
+    }
+
+    function extractTokens(input?: string): Set<string> {
+        const tokens = new Set<string>();
+        if (input === undefined) return tokens;
+        const regex = /{([^}]+)}/g;
+        let match;
+        while ((match = regex.exec(input)) !== null) {
+            tokens.add(match[1]);
+        }
+        return tokens;
+    }
 
     function close() {
         editingLabel.set(undefined);
+        saved = false;
+        suggestion = "";
     }
 
     function save() {
@@ -37,8 +54,14 @@
             busy = true;
             client
                 .setTranslationCorrection($locale, $editingLabel.key, suggestion)
-                .then(() => {
-                    editingLabel.set(undefined);
+                .then((success) => {
+                    if (success) {
+                        saved = true;
+                    } else {
+                        toastStore.showFailureToast(
+                            i18nKey("Sorry we were unable to save your suggestion"),
+                        );
+                    }
                 })
                 .finally(() => (busy = false));
         }
@@ -50,33 +73,58 @@
         <ModalContent on:close>
             <div class="header" slot="header">Suggest a translation correction</div>
             <div slot="body">
-                <p>
-                    The language you wish to submit a correction for is <span class="value"
-                        >{yourLanguage}</span>
-                </p>
-                <p>
-                    The English value is <span class="value"
-                        >{$_($editingLabel.key, { locale: "en" })}</span>
-                </p>
-                <p>The current translation is <span class="value">{$_($editingLabel.key)}</span></p>
-                {#if existingCorrection !== undefined}
+                {#if !saved}
                     <p>
-                        The current translation was provided by <span class="value"
-                            >{correctedBy}</span>
+                        The language you wish to submit a correction for is <span class="value"
+                            >{yourLanguage}</span>
                     </p>
+                    <p>
+                        The English value is <span class="value">{englishValue}</span>
+                    </p>
+                    <p>
+                        The current translation is <span class="value"
+                            >{$_($editingLabel.key)}</span>
+                    </p>
+                    <Legend label={i18nKey("Your proposed translation is")}></Legend>
+                    <TextArea
+                        minlength={1}
+                        maxlength={1000}
+                        disabled={busy}
+                        bind:value={suggestion}
+                        placeholder={i18nKey("Enter your suggestion")} />
+
+                    {#if suggestion !== "" && tokenMismatch}
+                        <ErrorMessage>
+                            Your suggested correction must contain the same &lbrace;tokens&rbrace;
+                            as the original English text
+                        </ErrorMessage>
+                    {/if}
+                {:else}
+                    <div class="saved">
+                        <p>
+                            Thank you for your suggestion. Please review the UI with your new
+                            suggestion in place. If you would like to make further changes just
+                            repeat this process until you are happy.
+                        </p>
+                        <p>
+                            Your suggestion will be reviewed by a platform operator and applied soon
+                            if it is approved.
+                        </p>
+                        <p>
+                            In the meantime it will appear locally for you unless / until you
+                            refresh the page.
+                        </p>
+                    </div>
                 {/if}
-                <Legend label={i18nKey("Your proposed translation is")}></Legend>
-                <TextArea
-                    minlength={1}
-                    maxlength={1000}
-                    disabled={busy}
-                    bind:value={suggestion}
-                    placeholder={i18nKey("Enter your suggestion")} />
             </div>
             <div slot="footer">
                 <ButtonGroup>
-                    <Button secondary on:click={close}>{"Cancel"}</Button>
-                    <Button on:click={save}>{"Save"}</Button>
+                    {#if saved}
+                        <Button on:click={close}>{"Close"}</Button>
+                    {:else}
+                        <Button secondary on:click={close}>{"Cancel"}</Button>
+                        <Button disabled={!valid} on:click={save}>{"Save"}</Button>
+                    {/if}
                 </ButtonGroup>
             </div>
         </ModalContent>