+++ frontend/openchat-client/src/openchat.ts
@@ -705,11 +705,6 @@ export class OpenChat extends OpenChatAgentWorker {
                         this.identityState.set({ kind: "registering" });
                         break;
                     case "created_user":
-                        console.log("About to get translation corrections");
-                        this.getTranslationCorrections().then((res) => {
-                            console.log("Corrections: ", res);
-                            applyTranslationCorrections(user.userId, res);
-                        });
                         this.onCreatedUser(user);
                         break;
                 }
@@ -5612,24 +5607,23 @@ export class OpenChat extends OpenChatAgentWorker {
         return community.localUserIndex;
     }
 
-    setTranslationCorrection(
-        locale: string,
-        key: string,
-        value: string,
-    ): Promise<TranslationCorrections> {
+    setTranslationCorrection(locale: string, key: string, value: string): Promise<boolean> {
+        const correction = {
+            locale,
+            key,
+            value,
+            proposedBy: this._liveState.user.userId,
+            proposedAt: Date.now(),
+            approved: false,
+        };
         return this.sendRequest({
             kind: "setTranslationCorrection",
-            correction: {
-                locale,
-                key,
-                value,
-                proposedBy: this._liveState.user.userId,
-                proposedAt: Date.now(),
-                approved: false,
-            },
-        }).then((updated) => {
-            applyTranslationCorrections(this._liveState.user.userId, updated);
-            return updated;
+            correction,
+        }).then((success) => {
+            if (success) {
+                applyTranslationCorrection(correction);
+            }
+            return success;
         });
     }
 
@@ -5645,9 +5639,6 @@ export class OpenChat extends OpenChatAgentWorker {
         return this.sendRequest({
             kind: "rejectTranslationCorrection",
             correction,
-        }).then((updated) => {
-            applyTranslationCorrections(this._liveState.user.userId, updated);
-            return updated;
         });
     }
 
@@ -5657,9 +5648,6 @@ export class OpenChat extends OpenChatAgentWorker {
         return this.sendRequest({
             kind: "approveTranslationCorrection",
             correction,
-        }).then((updated) => {
-            applyTranslationCorrections(this._liveState.user.userId, updated);
-            return updated;
         });
     }
 
@@ -6173,7 +6161,6 @@ export class OpenChat extends OpenChatAgentWorker {
     selectedMessageContext = selectedMessageContext;
     userGroupSummaries = userGroupSummaries;
     offlineStore = offlineStore;
-    translationCorrectionsStore = translationCorrectionsStore;
 
     // current community stores
     chatListScope = chatListScopeStore;